feat: Express版を削除しCloudflare Workers版に統一
- Express.jsの依存関係とファイルを完全に削除
- worker.tsをindex.tsにリネームして構造を統一
- Viteを導入してビルド環境を最新化
- Miniflareの設定を追加してローカル開発環境を最適化
- package.jsonのスクリプトを整理・簡素化
- wrangler.tomlに開発環境設定を追加
