+++ wrangler.toml
@@ -1,9 +1,21 @@
 name = "omikuji"
-main = "dist/worker.js"
+main = "src/index.ts"
 compatibility_date = "2024-01-01"
 
+# Development settings
+[dev]
+port = 8787
+local_protocol = "http"
+
+# Production environment
 [env.production]
 name = "omikuji"
 
+# Build settings
 [build]
-command = "npm run build:worker"
+command = "npm run build"
+
+# Miniflare settings for local development
+[miniflare]
+# Enable KV namespaces, Durable Objects, etc. when needed
+kv_persist = true
